photography_bot/requirements.txt
discord.py==2.5.2
Flask==3.1.1
python-dotenv==1.1.0